+++ domain/model/src/main/java/co/com/retowebflux/model/user/gateways/UserProviderGateway.java
@@ -4,5 +4,5 @@ import co.com.retowebflux.model.user.User;
 import reactor.core.publisher.Mono;
 
 public interface UserProviderGateway {
-    Mono<User> getUserById(Integer id);
+    Mono<User> getUserById(Long id);
 }
